+++ packages/e2e/test/integration/first.e2e.dart
@@ -8,19 +8,19 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint, type=warning
 import 'first.dart' as _i1;
-import 'a.dart' as _i1;
-import '../../lib/e2e/b.dart' as _i1;
-import 'c.dart' as _i1;
-import 'f.dart' as _i1;
+import 'a.dart' as _i2;
+import '../../lib/e2e/b.dart' as _i3;
+import 'c.dart' as _i4;
+import 'f.dart' as _i5;
 
 void simpleFromTestE2e([
   Object value = const _i1.Foo(),
 ]) =>
     throw UnimplementedError();
 void prefixToRelativePathE2e([
-  Object value = const _i1.A(),
-  Object value2 = const _i1.B(),
-  Object value3 = const _i1.C(),
-  Object value6 = const _i1.F(),
+  Object value = const _i2.A(),
+  Object value2 = const _i3.B(),
+  Object value3 = const _i4.C(),
+  Object value6 = const _i5.F(),
 ]) =>
     throw UnimplementedError();